# Fix parseInt overflow check false negative