--- a/bytes.go
+++ b/bytes.go
@@ -1,11 +1,8 @@
 package jsonparser
 
-import (
-	bio "bytes"
-)
-
-// minInt64 '-9223372036854775808' is the smallest representable number in int64
-const minInt64 = `9223372036854775808`
+const absMinInt64 = 1 << 63
+const maxInt64 = 1<<63 - 1
+const maxUint64 = 1<<64 - 1
 
 // About 2x faster then strconv.ParseInt because it only supports base 10, which is enough for JSON
 func parseInt(bytes []byte) (v int64, ok bool, overflow bool) {
@@ -19,29 +16,32 @@ func parseInt(bytes []byte) (v int64, ok bool, overflow bool) {
 		bytes = bytes[1:]
 	}
 
-	var b int64 = 0
+	var n uint64 = 0
 	for _, c := range bytes {
-		if c >= '0' && c <= '9' {
-			b = (10 * v) + int64(c-'0')
-		} else {
+		if c < '0' || c > '9' {
 			return 0, false, false
 		}
-		if overflow = (b < v); overflow {
-			break
+		if n > maxUint64/10 {
+			return 0, false, true
 		}
-		v = b
+		n *= 10
+		n1 := n + uint64(c-'0')
+		if n1 < n {
+			return 0, false, true
+		}
+		n = n1
 	}
 
-	if overflow {
-		if neg && bio.Equal(bytes, []byte(minInt64)) {
-			return b, true, false
+	if n > maxInt64 {
+		if neg && n == absMinInt64 {
+			return -absMinInt64, true, false
 		}
 		return 0, false, true
 	}
 
 	if neg {
-		return -v, true, false
+		return -int64(n), true, false
 	} else {
-		return v, true, false
+		return int64(n), true, false
 	}
 }